pyproject.toml
[tool.poetry]
name = "rb"
version = "1.0.0"
description = "Reebok"
authors = ["guimassoqueto <gui.massoqueto@gmail.com>"]
readme = "README.md"
packages = [
{include = "*.py"},
{include = "app/**/*.py"}
]

[tool.poetry.dependencies]
python = "^3.11"
pre-commit = "^3.4.0"
python-dotenv = "^1.0.0"
playwright = "^1.38.0"
selectolax = "^0.3.16"
aiohttp-retry = "^2.8.3"
pydantic = "^2.3.0"
fake-useragent = "^1.2.1"
psycopg = "^3.1.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
